version: 0.6.0.0
ADD: Added a searchbar on the general layout. (It shows on all pages)

ADD: added new parameters related to the search

ADD: Added an extension class for IEnumerable<FlashCardModel> so its possible to filter by topic and / or subtopic.  (Default is languages / japanese)

FIX: Fixed the issue with not being able to edit or add front / back content.

NOTE: This is probably going to be the last update or second to last update, before upping the version to 1.0.0.0 aka. full release.